Split app concerns
