+++ Makefile
@@ -1,0 +1,5 @@
+help:
+	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-30s\033[0m %s\n", $$1, $$2}'
+
+duck: ## Make a duck noise
+	@echo Quack quack!
