+++ site/README.md
@@ -1,0 +1,5 @@
+Site
+====
+
+This is the code responsible for the compilation of the site. It's a load of
+React templates and CSS. Rock on.
